feature,pricing,location icin api controller lar olusturuldu ve db ye örnek veri girisi yapıldı

+++ DtoLayer/CarDtos/ResultCarWithBrandsDto.cs
@@ -16,7 +16,7 @@ namespace DtoLayer.CarDtos
         public int Km { get; set; }
         public string Transmission { get; set; }
         public byte Seat { get; set; }
-        public byte Luggage { get; set; }
+        public int Luggage { get; set; }
         public string Fuel { get; set; }
         public string BigImageUrl { get; set; }
     }
