+++ EntitityLayer/Entities/Car.cs
@@ -16,7 +16,7 @@ namespace EntitityLayer.Entities
         public int Km { get; set; }
         public string Transmission { get; set; }
         public byte Seat { get; set; }
-        public byte Luggage { get; set; }
+        public int Luggage { get; set; }
         public string Fuel { get; set; }
         public string BigImageUrl { get; set; }
         public List<CarFeature> CarFeatures { get; set; }
